requirements.txt
apache-airflow[gcp,crypto]
httplib2
jsonschema
pyyaml